+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Starting with Peertube 3.2.0, builtin prosody save room history on server. So when a user connects, he can get previously send messages.
 * Starting with Peertube 3.2.0, builtin prosody also activate mod_muc_moderation, enabling moderators to moderate messages.
 * Prosody log level will be the same as the Peertube's one.
+* Prosody log rotation every 24 hour.
 
 ### Fixes